nixos-option: rewrite as a nix script
This ports the functionality of the C++ nixos-option to a nix script (with a
tiny shellscript for argument processing and invoking the nix script).
Benefits compared to the native binary include no longer being tied to a
specific nix version, generally improved maintainability and improved
stability.

The main tradeoff is that the C++ version would have better access to
introspecting and reporting errors nicely, but that doesn't seem to have been
the case in practice anyway.  The other tradeoff is that we generate all the
output at the end instead of streaming it as we traverse the option tree.